corrections erreurs

+++ bot/commands/moderator/Purge.js
@@ -15,13 +15,13 @@ module.exports = {
     message.channel.bulkDelete(amount).catch(async err => {
       let msg = await message.channel.send(':x: Due to Discord Limitations, I cannot delete messages older than 14 days')
       setTimeout(() => {
-        msg.delete()
+        if(msg) msg.delete()
       }, 2000)
     })
 
     let msg = await message.channel.send(`Deleted \`${amount}\` messages`)
     setTimeout(() => {
-        msg.delete()
+        if(msg) msg.delete()
     }, 2000)
   },
 }
